Make session credential cache naming a bit smarter
if the config profile specifies a source profile, create the cache
file using that name (so profiles using the same source profile can
use a common cache).  Otherwise create the cache using the profile
name, which makes the cache file specific to the profile.

For cases where neither are set (ex. role arn passed directly on the
command line), the file name will be .aws_session_token_, but for the
most part these cases should be pretty limited in scope.