Add mock EC2 metadata service provider
Create a mock endpoint to act as an EC2 instance profile credential
provider.  This should allow aws-runas to be run in the background to
answer SDK requests for credentials from the EC2 metadata service
(which is typically in the default SDK credential lookup chain).  The
credentials returned will have the permissions applicable to the profile
or role ARN which aws-runas was set to use at startup.