Handle failure for IAM user lookup better
If initially run with role credentials in the environment, the call to
iam.GetUser() fails, since the credentials don't map to an actual IAM
user.  Attempt to handle this case a bit more gracefully by retrying the
iam.GetUser() call with the AWS credential environment variables unset,
so it will try other credential sources for the actual user credentials.

Fixes #21