Use linux env var setting syntax if SHELL env var set on Windows
SHELL env var is not set on Windows platform when using cmd.exe or
PowerShell, and env vars can be set using the usual Windows 'set'
syntax.  However, if SHELL is set on Windows, we may be running in a
session that emulates linux (git bash, or Cygwin), and we should use
linux env var syntax instead.