Drop process privileges as soon as we're done with the "admin-level" stuff
After setting up the network interface and port 80 listening socket,
change the process uid and gid back to the value of the user who started
us. In addition to a good security practice, this also keeps the
ownership of the files in the credentials cache sane, otherwise it would
create all the files as root, which is probably not what we want.

So far this only works on unix-like systems, fingers crossed that things
don't have to be changed for Windows

New call to UserHomeDir() makes Go 1.12 a requirement for building