Allow for auto-refresh of role credentials when wrapping a command
When using aws-runas to execute a command, automatically track the
expiration of the role credentials and refresh them once they enter the
expiry window.

The necessitates a change from setting credentials as env vars before
executing the process to writing the role creds to a temporary credential
file and setting the appropriate env var to have the spawned process use
that credential file instead of the default one.