Allow for auto-refresh of role credentials when wrapping a command
Instead of using a temporary credential file, start an HTTP server which
mimics the ECS credential endpoint, and modify the environment such that
the called command will reference the endpoint for credentials.  This
should allow the command to automatically refresh the role credentials
while the process executes, for as long as the session token credentials
are valid.