add storage/retrieval of saml password from credentials file
Add the 'password' sub-command to obfuscate and store the password in the .aws/credentials file
using the SAML endpoint URL as the section/profile name and a custom attribute name.

Update the SAML credential gathering logic to reference this information if the env vars or
command line option to pass the password are missing.

Fixes #45