smash some bugs in the metadata service
make the startup message for the metadata service more clear that the 1st
thing you should do is access the web interface and select a role.

make sure that the saml cookies file has appropriate ownership if it gets
created during the initialization of the metadata service.

pass the saml.AwsClient from main into the metadata options so that we have
a non-nil client at the start, if we need one.