refactor OneLogin API interactions so we get persistent user logins
Previously, interacting directly with the SAML endpoint, you needed to provide the user
credentials every time you called aws-runas, which would trigger a reprompt for MFA entry.
That is very annoying.  Now we log in the user to OneLogin so we can get the required HTTP
cookie to persist the login state, and use a different URL for obtaining the SAML data.

The new URL is in the general form of:

https://my-tenant.onelogin.com/trust/saml2/launch/__app-id__

where the app-id value can be found on the user's application landing page, hovering over
the OneLogin AWS Application, and getting the last element in the URL path.