Add feature to allow user selection of mfa type
When configured for SAML authentication, allow user to
define the mfa type to use.  This will override the "auto"
MFA processing logic in the external provider and only
use the requested type.  Behavior is undefined if the
requested type is not available for the account (expect an
error at the very least, and possibly a panic())

Configuration can be specified using the mfa_type config
file attribute, the -t command line option, or the
MFA_TYPE environment variable. Standard precedence rules
apply.

Also made a metric pantload of local lint check updates
after addinng the 'lint' target to the Makefile