Always merge a valid config object when using the config chain loader.
This allows the use of external IdP configuration in the default profile
so a role ARN can be passed to the command instead of only a profile name.
Previously the config file info was not being used because the role ARN
was not resolved to a 'valid' profile name in the config file, and was skipped.